+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/DependencyGraphNodeCache.kt
@@ -74,7 +74,7 @@ internal class DependencyGraphNodeCache(
           bindingStack,
           parentTracer,
           metroGraph,
-          dependencyGraphAnno
+          dependencyGraphAnno,
         )
       }
     }
@@ -111,8 +111,10 @@ internal class DependencyGraphNodeCache(
     private val injectors = mutableListOf<Pair<MetroSimpleFunction, IrContextualTypeKey>>()
     private val includedGraphNodes = mutableMapOf<IrTypeKey, DependencyGraphNode>()
     private val graphTypeKey = IrTypeKey(graphDeclaration.typeWith())
+    private val sourceGraphTypeKey = IrTypeKey(graphDeclaration.sourceGraphIfMetroGraph.typeWith())
     private val graphContextKey = IrContextualTypeKey.create(graphTypeKey)
     private val bindingContainers = mutableSetOf<BindingContainer>()
+    private val managedBindingContainers = mutableSetOf<IrClass>()
 
     private val dependencyGraphAnno =
       cachedDependencyGraphAnno
@@ -214,6 +216,7 @@ internal class DependencyGraphNodeCache(
         nonNullCreator.parameters.regularParameters.forEachIndexed { i, parameter ->
           if (parameter.isBindsInstance) return@forEachIndexed
 
+          // It's an `@Includes` parameter
           val klass = parameter.typeKey.type.rawType()
           val sourceGraph = klass.sourceGraphIfMetroGraph
 
@@ -221,11 +224,20 @@ internal class DependencyGraphNodeCache(
 
           // Add any included graph provider factories IFF it's a binding container
           if (nonNullCreator.bindingContainersParameterIndices.isSet(i)) {
-            val bindingContainer =
-              bindingContainerTransformer.findContainer(sourceGraph)
-                ?: error("Binding container not found for type ${sourceGraph.classId}")
-
-            bindingContainers += bindingContainer
+            // Include the container itself and all its transitively included containers
+            val allContainers =
+              bindingContainerTransformer.resolveAllBindingContainersCached(setOf(sourceGraph))
+            bindingContainers += allContainers
+            // Track which transitively included containers be managed
+            for (container in allContainers) {
+              if (container.ir == klass) {
+                // Don't mark the parameter class itself as managed since we're taking it as an
+                // input
+                continue
+              } else if (container.canBeManaged) {
+                managedBindingContainers += container.ir
+              }
+            }
             return@forEachIndexed
           }
 
@@ -379,9 +391,8 @@ internal class DependencyGraphNodeCache(
               val rawType = metroFunction.ir.returnType.rawType()
               val functionParent = rawType.parentClassOrNull
 
-              val isGraphExtensionFactory = rawType.isAnnotatedWithAny(
-                symbols.classIds.graphExtensionFactoryAnnotations
-              )
+              val isGraphExtensionFactory =
+                rawType.isAnnotatedWithAny(symbols.classIds.graphExtensionFactoryAnnotations)
 
               if (isGraphExtensionFactory) {
                 // For factories, add them to accessors so they participate in the binding graph
@@ -392,16 +403,22 @@ internal class DependencyGraphNodeCache(
                 val samMethod = rawType.singleAbstractFunction()
                 val graphExtensionType = samMethod.returnType
                 val graphExtensionTypeKey = IrTypeKey(graphExtensionType)
-                graphExtensions.getOrPut(graphExtensionTypeKey, ::mutableListOf) += GraphExtensionAccessor(
-                  accessor = metroFunction,
-                  key = factoryContextKey,
-                  isFactory = true,
-                  isFactorySAM = false,
-                )
+                if (graphExtensionTypeKey != sourceGraphTypeKey) {
+                  // Only add it to our graph extensions if it's not exposing itself
+                  graphExtensions.getOrPut(graphExtensionTypeKey, ::mutableListOf) +=
+                    GraphExtensionAccessor(
+                      accessor = metroFunction,
+                      key = factoryContextKey,
+                      isFactory = true,
+                      isFactorySAM = false,
+                    )
+                }
               } else {
                 // Regular graph extension
-                val isSamFunction = metroFunction.ir.overriddenSymbolsSequence()
-                  .any { it.owner.parentClassOrNull?.classId in graphExtensionSupertypes }
+                val isSamFunction =
+                  metroFunction.ir.overriddenSymbolsSequence().any {
+                    it.owner.parentClassOrNull?.classId in graphExtensionSupertypes
+                  }
 
                 val contextKey =
                   if (
@@ -414,12 +431,13 @@ internal class DependencyGraphNodeCache(
                   } else {
                     IrContextualTypeKey.from(declaration)
                   }
-                graphExtensions.getOrPut(contextKey.typeKey, ::mutableListOf) += GraphExtensionAccessor(
-                  metroFunction,
-                  key = contextKey,
-                  isFactory = false,
-                  isFactorySAM = isSamFunction,
-                )
+                graphExtensions.getOrPut(contextKey.typeKey, ::mutableListOf) +=
+                  GraphExtensionAccessor(
+                    metroFunction,
+                    key = contextKey,
+                    isFactory = false,
+                    isFactorySAM = isSamFunction,
+                  )
               }
               hasGraphExtensions = true
             } else if (isInjector) {
@@ -450,9 +468,8 @@ internal class DependencyGraphNodeCache(
             val getter = declaration.getter!!
 
             val rawType = getter.returnType.rawType()
-            val isGraphExtensionFactory = rawType.isAnnotatedWithAny(
-              symbols.classIds.graphExtensionFactoryAnnotations
-            )
+            val isGraphExtensionFactory =
+              rawType.isAnnotatedWithAny(symbols.classIds.graphExtensionFactoryAnnotations)
             var isGraphExtension = isGraphExtensionFactory
 
             // If the overridden symbol has a default getter/value then skip
@@ -490,16 +507,22 @@ internal class DependencyGraphNodeCache(
                 val samMethod = rawType.singleAbstractFunction()
                 val graphExtensionType = samMethod.returnType
                 val graphExtensionTypeKey = IrTypeKey(graphExtensionType)
-                graphExtensions.getOrPut(graphExtensionTypeKey, ::mutableListOf) += GraphExtensionAccessor(
-                  metroFunction,
-                  key = contextKey,
-                  isFactory = true,
-                  isFactorySAM = false,
-                )
+                if (graphExtensionTypeKey != sourceGraphTypeKey) {
+                  // Only add it to our graph extensions if it's not exposing itself
+                  graphExtensions.getOrPut(graphExtensionTypeKey, ::mutableListOf) +=
+                    GraphExtensionAccessor(
+                      metroFunction,
+                      key = contextKey,
+                      isFactory = true,
+                      isFactorySAM = false,
+                    )
+                }
               } else {
                 // Regular graph extension
-                val isSamFunction = metroFunction.ir.overriddenSymbolsSequence()
-                  .any { it.owner.parentClassOrNull?.classId in graphExtensionSupertypes }
+                val isSamFunction =
+                  metroFunction.ir.overriddenSymbolsSequence().any {
+                    it.owner.parentClassOrNull?.classId in graphExtensionSupertypes
+                  }
                 val functionParent = rawType.parentClassOrNull
                 val finalContextKey =
                   if (
@@ -512,12 +535,13 @@ internal class DependencyGraphNodeCache(
                   } else {
                     contextKey
                   }
-                graphExtensions.getOrPut(finalContextKey.typeKey, ::mutableListOf) += GraphExtensionAccessor(
-                  metroFunction,
-                  key = finalContextKey,
-                  isFactory = false,
-                  isFactorySAM = isSamFunction,
-                )
+                graphExtensions.getOrPut(finalContextKey.typeKey, ::mutableListOf) +=
+                  GraphExtensionAccessor(
+                    metroFunction,
+                    key = finalContextKey,
+                    isFactory = false,
+                    isFactorySAM = isSamFunction,
+                  )
               }
               hasGraphExtensions = true
             } else {
@@ -554,7 +578,6 @@ internal class DependencyGraphNodeCache(
         extendedGraphNodes[node.typeKey] = node
       }
 
-      val managedBindingContainers = mutableSetOf<IrClass>()
       bindingContainers +=
         dependencyGraphAnno
           ?.bindingContainerClasses(includeModulesArg = options.enableDaggerRuntimeInterop)