+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/transformers/InjectConstructorTransformer.kt
@@ -28,6 +28,7 @@ import dev.zacsweers.metro.compiler.ir.requireSimpleFunction
 import dev.zacsweers.metro.compiler.ir.thisReceiverOrFail
 import dev.zacsweers.metro.compiler.ir.trackFunctionCall
 import dev.zacsweers.metro.compiler.ir.typeAsProviderArgument
+import dev.zacsweers.metro.compiler.reportCompilerBug
 import java.util.Optional
 import kotlin.jvm.optionals.getOrNull
 import org.jetbrains.kotlin.descriptors.ClassKind
@@ -152,7 +153,7 @@ internal class InjectConstructorTransformer(
         generatedFactories[injectedClassId] = Optional.empty()
         return null
       } else {
-        error(
+        reportCompilerBug(
           "No expected factory class generated for '${declaration.kotlinFqName}'. Report this bug with a repro case at https://github.com/zacsweers/metro/issues/new"
         )
       }
@@ -297,7 +298,7 @@ internal class InjectConstructorTransformer(
                 functionParamsByName.getValue(targetParam.originalName)
               }
 
-              else -> error("Unmatched top level injected function param: $targetParam")
+              else -> reportCompilerBug("Unmatched top level injected function param: $targetParam")
             }
           }
 
@@ -411,7 +412,7 @@ internal class InjectConstructorTransformer(
                     functionParamsByName.getValue(targetParam.originalName)
                   }
 
-                  else -> error("Unmatched top level injected function param: $targetParam")
+                  else -> reportCompilerBug("Unmatched top level injected function param: $targetParam")
                 }
               }