+++ CHANGELOG.md
@@ -4,6 +4,28 @@ Changelog
 **Unreleased**
 --------------
 
+0.6.3
+-----
+
+_2025-08-23_
+
+* **Enhancement**: Allow `@Includes` parameters that are binding containers to transitively include other binding containers.
+* **Fix**: Ensure provider fields for graph instances when needed by extensions.
+
+Special thanks to [@hossain-khan](https://github.com/hossain-khan), [@vRallev](https://github.com/vRallev), and [@erawhctim](https://github.com/erawhctim) for contributing to this release!
+
+0.6.2
+-----
+
+_2025-08-22_
+
+* **Enhancement**: Add diagnostic for contributed binding containers with no (visible) no-arg constructor.
+* **Enhancement**: Add hint for missing bindings if the binding appears to be contributed by an `internal` type in another, non-friend module.
+* **Fix**: Don't allocate new fields for deferred bindings reserved by extension graphs.
+* **Fix**: Allow graph extensions to expose their own factories if inherited from parents.
+
+Special thanks to [@KevinGuitar](https://github.com/KevinGuitar), [@hossain-khan](https://github.com/@hossain-khan) and [@ChrisBanes](https://github.com/ChrisBanes) for contributing to this release!
+
 0.6.1
 -----